requirements.txt
cloudshell-automation-api
ldap3
dacite
pywin32